Create enemyprojectileexample.lua
This is a short example of a local Runtime listener that is generated with each spawning of an enemy guard robot. There's a lot going on here that I don't think needs to be explained, mostly because it's only germane to my project. Check out my "hascollided" repo for more info on that "hasCollided" functionality.